+++ abstractions/python/kiota/abstractions/authentication/allowed_hosts_validator.py
@@ -14,6 +14,9 @@ class AllowedHostsValidator:
             allowed_hosts (List[str]): A list of valid hosts.  If the list is empty, all hosts
             are valid.
         """
+        if not isinstance(allowed_hosts, list):
+            raise TypeError("Allowed hosts must be a list of strings")
+
         self.allowed_hosts: Set[str] = {x.lower() for x in allowed_hosts}
 
     def get_allowed_hosts(self) -> List[str]:
@@ -31,6 +34,8 @@ class AllowedHostsValidator:
             allowed_hosts (List[str]): A list of valid hosts.  If the list is empty, all hosts
             are valid
         """
+        if not isinstance(allowed_hosts, list):
+            raise TypeError("Allowed hosts must be a list of strings")
         self.allowed_hosts = {x.lower() for x in allowed_hosts}
 
     def is_url_host_valid(self, url: str) -> bool: