+++ authentication/python/azure/authentication_azure/azure_identity_access_token_provider.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Dict, List, Optional, Set
+from typing import TYPE_CHECKING, Dict, List, Optional
 from urllib.parse import urlparse
 
 from kiota.abstractions.authentication import AccessTokenProvider, AllowedHostsValidator
@@ -16,10 +16,10 @@ class AzureIdentityAccessTokenProvider(AccessTokenProvider):
         credentials: "AsyncTokenCredential",
         options: Optional[Dict],
         scopes: List[str] = ['https://graph.microsoft.com/.default'],
-        allowed_hosts: Set[str] = {
+        allowed_hosts: List[str] = [
             'graph.microsoft.com', 'graph.microsoft.us', 'dod-graph.microsoft.us',
             'graph.microsoft.de', 'microsoftgraph.chinacloudapi.cn', 'canary.graph.microsoft.com'
-        },
+        ],
     ) -> None:
         if not credentials:
             raise Exception("Parameter credentials cannot be null")