+++ authentication/python/azure/tests/helpers.py
@@ -1,0 +1,14 @@
+from dataclasses import dataclass
+
+@dataclass
+class DummyToken:
+    token: str
+        
+class DummyAzureTokenCredential():
+    
+    async def get_token(*args):
+        return DummyToken(token="This is a dummy token")
+    
+    async def close(*args):
+        pass
+