+++ serialization/python/json/serialization_json/json_parse_node_factory.py
@@ -35,4 +35,5 @@ class JsonParseNodeFactory(ParseNodeFactory):
             raise TypeError("Content cannot be null")
 
         content_as_str = content.decode('utf-8')
-        return JsonParseNode(content_as_str)
+        content_dict = json.loads(content_as_str)
+        return JsonParseNode(content_dict)